packages/cheese3d/pyproject.toml
[project]
name = "cheese3d"
version = "0.1.0"
description = "Sensitive detection and analysis of whole-face movement in mice."
authors = [{ name = "Hou Lab CSHL", email = "the.hou.lab@gmail.com" }]
readme = "../../README.md"
requires-python = "==3.10.*"
scripts = { cheese3d = "cheese3d.cli:cli" }
dependencies = [
    "pip>=25.2,<26",
    "ipython",
    "ipykernel",
    "jupyter",
    "protobuf < 3.20",
    "tables == 3.7.*",
    "pandas == 2.1.*",
    "configobj == 5.0.8",
    "hydra-core",
    "joblib",
    "torch <= 1.12.0",
    "aniposelib @ git+https://github.com/Hou-Lab-CSHL/aniposelib.git@50c5b9d",
    "anipose @ git+https://github.com/Hou-Lab-CSHL/anipose.git",
    "opencv-python == 4.7.*",
    "opencv-python-headless == 4.7.*",
    "opencv-contrib-python == 4.7.*",
    "pims == 0.6.*",
    "seaborn",
    "open-ephys-python-tools",
    "typer-slim[standard]>=0.15.3,<0.16",
    "questionary",
    "textual",
    "textual-fspicker",
    "textual_serve",
    "cheese3d_annotator == 0.1.0",

    # Linux-specific deps
    "blosc2; sys_platform == 'linux'",
    "tensorflow == 2.10.*; sys_platform == 'linux'",
    "jax[cuda12_pip] >=0.4,<0.5; sys_platform == 'linux'",
    "deeplabcut[tf] @ git+https://github.com/Hou-Lab-CSHL/DeepLabCut.git ; sys_platform == 'linux'",

    # Apple silicon-specific deps
    "tensorflow-macos == 2.10.*; sys_platform == 'darwin'",
    "tensorflow-metal == 0.6.*; sys_platform == 'darwin'",
    "jax[cpu] >=0.4,<0.5; sys_platform == 'darwin'",
    "deeplabcut[apple_mchips] @ git+https://github.com/Hou-Lab-CSHL/DeepLabCut.git ; sys_platform == 'darwin'",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.metadata]
allow-direct-references = true
